renamed exportModal and locationSuggester files to be consistent
